requirements.txt
locustio==0.7.5
coveralls==1.1
docopt==0.6.2
flask==0.11.1
flask-cors==3.0.2
kafka-python==1.3.1
mock==1.0.1